Update enemy graphics and drawing method
In this update, new enemy types ('Golem', 'Sprinter', and 'Bombenschiff') are introduced with corresponding BufferedImage and File instances. These images are read and handled within a 'try' block of the code; and the switch condition in the 'paint()' method has been modified to draw these enemies based on their type. Furthermore, some background image file extensions were updated for consistency.