Refactored code and updated game logic
The Main.java class was refactored, removing redundant variables and updating the game loop logic. Simultaneously, improved the way buildings are chosen and cleared in HauptgrafikSpiel.java, and updated the process of removing buildings. The method to attack and the object's lifetime computation were also revised. These changes make code easier to understand and improve the game dynamic.